pyproject.toml
[tool.poetry]
name = "aiogram-nats"
version = "0.1.0"
description = ""
authors = ["NezBut <aleksejstacaenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.10.0"
aiogram-dialog = "^2.1.0"
dishka = "^1.2.0"
structlog = "^24.4.0"
nats-py = "^2.8.0"
dynaconf = "^3.2.6"
taskiq = "^0.11.6"
taskiq-nats = "^0.4.0"
taskiq-redis = "^1.0.0"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
adaptix = "3.0.0b7"
fluentogram = "^1.1.7"
asyncpg = "^0.29.0"
uvicorn = "^0.30.6"
fastapi = {extras = ["standard"], version = "^0.115.0"}
asgi-monitor = "^0.5.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
mypy = "^1.11.1"
ruff = "^0.5.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true

[tool.ruff]
src = ["aiogram_nats"]
line-length = 160
target-version = "py311"
format.indent-style = "space"
indent-width = 4

include = ["aiogram_nats/**.py", "tests/**.py"]

lint.select = [
    "ALL"
]
lint.ignore = [
    "ARG",
    "ANN",
    "D212",
    "D400",
    "D401",
    "D415",
    "D210",
    "D104",
    "D100",
    "D107",
    "EM101",
    "EM102",
    "PT001",
    "PT023",
    "SIM108",
    "SIM114",
    "TRY003",
    "PLW2901",
    "RET505",
    "PLR0913",
    "UP038",
    "TCH001",
    "RUF001",
    "RUF002",
    "RUF003",
    "G004",
    "INP001",
    "UP007",
    "PYI013",
    "PIE790",
    "D211",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["TID252", "PLR2004", "S101"]

[tool.ruff.lint.isort]
no-lines-before = ["local-folder"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "parents"
